Add ui package (Tiles, GameLog); add new Inventory class
Tiles has simply been moved (and referenced accordingly)
GameLog subtracts from Main the log ScrollPane and its child
TextArea along with the critical pushInLog(String) method in a
ui class of its own.

Some of the game logic associated to inventory has been also moved
in a similar fashion to the Inventory class in the logic package
(note that the toString method is called on the Inventory object,
not its List<Item> from getList)

Both GameLog and Inventory are singletons.